requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
Django==4.2.3
djangorestframework==3.14.0
Flask==2.3.2
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
magic-filter==1.0.10
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
